Move stale PR detection logic from calculator.py to _stale_pr.py
Co-locate StalePr dataclass with its detection functions.
calculator.py drops from 247 to 180 lines.
